+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 - *(explorer_card)* Created the better card values img to return a example of view
 - *(attachment)* Improved a minimal design on cards
 - *(items)* Load items from localhost api
+- *(items)* Load error report email if fails to load cards
 
 ### ⚙️ Miscellaneous Tasks